Moving prebuilt rules tests and adjusting retry logic
